Split up test-cases into different files based on functionality
Issue: #13
